refactor: / 로 접근을 로그인을 취급
